+++ API/CopilotService/__init__.py
@@ -11,7 +11,7 @@ from cli_validator.result import CommandSource
 from common import validate_command_in_task
 from common.auth import get_auth_token_for_learn_knowlegde_index, verify_token
 from common.correct import correct_scenario
-from common.exception import CopilotException, GPTInvalidResultException, ParameterException
+from common.exception import CopilotException, GPTInvalidResultException, ParameterException, KnowledgeSearchException
 from common.prompt import DEFAULT_GENERATE_SCENARIO_MSG, DEFAULT_SPLIT_TASK_MSG
 from common.service_impl.chatgpt import gpt_generate, num_tokens_from_message
 from common.service_impl.knowledge_base import knowledge_search, pass_verification
@@ -21,7 +21,8 @@ from common.service_impl.learn_knowledge_index import (filter_chunks_by_keyword_
                                                        retrieve_chunks_for_atomic_task,
                                                        trim_command_and_chunk_with_invalid_params)
 from common.telemetry import telemetry
-from common.util import generate_response, get_param, get_param_enum, get_param_int, get_param_str
+from common.util import generate_response
+from common.param import get_param, get_param_enum, get_param_int, get_param_str
 
 logger = logging.getLogger(__name__)
 
@@ -41,55 +42,66 @@ def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
         top_num = get_param_int(req, 'top_num', default=5)
         service_type = get_param_enum(req, 'type', ServiceType, default=os.environ.get("DEFAULT_SERVICE_TYPE", ServiceType.GPT_GENERATION))
     except ParameterException as e:
-        logger.error(f'Response Status 400: ParameterException: {e.msg}')
+        logger.error(f'Response Status 400: ParameterException: {e.msg}', exc_info=e)
         return func.HttpResponse(e.msg, status_code=400)
+
+    try:
+        result = copilot_service(context, question, history, top_num, service_type)
+    except CopilotException as e:
+        logger.error(f'Error: CopilotException: {e.msg}', exc_info=e)
+        return func.HttpResponse(e.to_response(), status_code=200)
+    except Exception as e:
+        logger.error(f'Unexpected Error: {str(e)}', exc_info=e)
+        return func.HttpResponse(generate_response(
+            {}, 500,
+            'We encountered an unexpected error while processing your request. '
+            'Please try again later or contact our support team for assistance.'))
+    return func.HttpResponse(generate_response(result, 200))
+
+
+def copilot_service(context, question, history, top_num=5, service_type=ServiceType.GPT_GENERATION):
 
     system_msg = os.environ.get("OPENAI_GENERATE_SCENARIO_MSG", default=DEFAULT_GENERATE_SCENARIO_MSG)
 
-    try:
-        result = []
-        if service_type == ServiceType.KNOWLEDGE_SEARCH:
-            try:
-                result = knowledge_search(question, top_num)
-                if len(result) == 0 or not pass_verification(context, question, result):
-                    result = []
-            except HttpResponseError as e:
-                logger.error('Response Status 500: Error from knowledge search: \n%s', e, exc_info=e)
-                return func.HttpResponse(f'Knowledge Search failed: {e.message}', status_code=500)
+    result = []
+    if service_type == ServiceType.KNOWLEDGE_SEARCH:
+        try:
+            result = knowledge_search(question, top_num)
+            if len(result) == 0 or not pass_verification(context, question, result):
+                result = []
+        except HttpResponseError as e:
+            raise KnowledgeSearchException() from e
 
-            return func.HttpResponse(generate_response(result, 200))
+        return result
 
-        if os.environ.get('ENABLE_RETRIEVAL_AUGMENTED_GENERATION', "true").lower() == "true":
-            task_list, usage_context = asyncio.run(_retrieve_context_from_learn_knowledge_index(context, question))
-            token_limit = int(os.environ.get("CONTEXT_TOKEN_LIMIT", 4096))
-            completion_tokens = int(os.environ.get('OPENAI_MAX_TOKENS', 4000))   # The default value should be the same as the one in initialize_chatgpt_service_params
-            factor = float(os.environ.get('ESTIMATION_ADJUSTMENT_FACTOR', 0.95))
-            system_msg_tokens = num_tokens_from_message(system_msg) or 0
-            token_remains = (token_limit - completion_tokens) * factor - system_msg_tokens
-            question = _add_context_to_queston(context, question, task_list, usage_context, token_limit=token_remains)
+    if os.environ.get('ENABLE_RETRIEVAL_AUGMENTED_GENERATION', "true").lower() == "true":
+        task_list, usage_context = asyncio.run(_retrieve_context_from_learn_knowledge_index(context, question))
+        token_limit = int(os.environ.get("CONTEXT_TOKEN_LIMIT", 4096))
+        completion_tokens = int(os.environ.get('OPENAI_MAX_TOKENS', 4000))   # The default value should be the same as the one in initialize_chatgpt_service_params
+        factor = float(os.environ.get('ESTIMATION_ADJUSTMENT_FACTOR', 0.95))
+        system_msg_tokens = num_tokens_from_message(system_msg) or 0
+        token_remains = (token_limit - completion_tokens) * factor - system_msg_tokens
+        question = _add_context_to_queston(context, question, task_list, usage_context, token_limit=token_remains)
 
-        if service_type == ServiceType.GPT_GENERATION:
+    if service_type == ServiceType.GPT_GENERATION:
+        context.custom_context.gpt_task_name = 'GENERATE_SCENARIO'
+        gpt_result = gpt_generate(context, system_msg, question, history)
+        result = [_build_scenario_response(gpt_result)] if gpt_result else []
+
+    elif service_type == ServiceType.MIX:
+        try:
+            result = knowledge_search(question, top_num)
+        except HttpResponseError as e:
+            logger.error('Error from knowledge search: \n%s', e, exc_info=e)
+            result = []
+
+        if len(result) == 0 or not pass_verification(context, question, result):
             context.custom_context.gpt_task_name = 'GENERATE_SCENARIO'
             gpt_result = gpt_generate(context, system_msg, question, history)
             result = [_build_scenario_response(gpt_result)] if gpt_result else []
 
-        elif service_type == ServiceType.MIX:
-            try:
-                result = knowledge_search(question, top_num)
-            except HttpResponseError as e:
-                logger.error('Error from knowledge search: \n%s', e, exc_info=e)
-                result = []
-
-            if len(result) == 0 or not pass_verification(context, question, result):
-                context.custom_context.gpt_task_name = 'GENERATE_SCENARIO'
-                gpt_result = gpt_generate(context, system_msg, question, history)
-                result = [_build_scenario_response(gpt_result)] if gpt_result else []
-
-        result = [correct_scenario(s) for s in result]
-    except CopilotException as e:
-        logger.error(f'Response Status 500: CopilotException: {e.msg}')
-        return func.HttpResponse(e.msg, status_code=500)
-    return func.HttpResponse(generate_response(result, 200))
+    result = [correct_scenario(s) for s in result]
+    return result
 
 
 async def _retrieve_context_from_learn_knowledge_index(context, question):
